Module rdiff::string_diff

Used for finding the minimal set of operations to transform one string into another.

The primary function of this module is find diff.

Structs

EditDistance
Used as the classiscal definition of edit distance.

Traits

OperationScore
Used to calculate the score for each operation that will be performed. The score can be static, or it can vary based on which character is being deleted inserted or substituted. It is highly recommended to inline the implementation of these characters

Functions

find_diff
Finds the difference on a character by character level between two strings